refactor : LikeUserInfo를 내부클래스 -> 외부클래스로 분리 (부적절)
